requirements.txt
numpy==1.24.0
onnxruntime==1.14.1
opencv_python==4.6.0.66
Pillow==9.4.0
requests==2.28.1
torch==1.13.1
torchvision==0.14.1
banana-dev